change uint64_t to uint32_t where possible

+++ KDNode.cpp
@@ -8,8 +8,8 @@ KDNode::KDNode(float median) {
     this->val.median = median;
 }
 
-KDNode::KDNode(uint64_t start, uint64_t end) {
-    this->left = (KDNode *)start;
-    this->right = (KDNode *)end;
+KDNode::KDNode(uint32_t start, uint32_t end) {
+    this->left = (KDNode *)((uint64_t)start);
+    this->right = (KDNode *)((uint64_t)end);
     this->val.isLeaf = LEAF_VAL;
 }
