+++ KDTree.h
@@ -23,24 +23,24 @@ struct CompareDistance {
 
 class KDTree {
 public:
-    KDTree(float *points, uint64_t nPoints, uint64_t nDim, int nCores);
+    KDTree(float *points, uint32_t nPoints, uint32_t nDim, int nCores);
     ~KDTree();
-    void query(float *queries, uint64_t nQueries, uint64_t k, float *out, int nCores);
+    void query(float *queries, uint32_t nQueries, uint32_t k, float *out, int nCores);
     void print(void);
 private:
     KDNode *root;
     float *points;
-    uint64_t nPoints, nDim;
+    uint32_t nPoints, nDim;
     float distanceToPoint(float *point1, float *point2);
-    void queryHelper(float *queries, uint64_t nQueries, uint64_t k, float *out);
+    void queryHelper(float *queries, uint32_t nQueries, uint32_t k, float *out);
     void getNN(KDNode *node,
                std::priority_queue<struct pair, std::vector<struct pair>, CompareDistance> &nn,
                float *queryPoint,
                int currD);
     void printVisitor(KDNode *node, int currD);
-    float getPivot(uint64_t startIndex, uint64_t endIndex, uint64_t currd);
-    uint64_t partition(uint64_t startIndex, uint64_t endIndex, uint64_t currd, float *pivotVal);
-    void buildTree(KDNode **node, uint64_t startIndex, uint64_t endIndex, uint64_t currd, int nCores);
+    float getPivot(uint32_t startIndex, uint32_t endIndex, uint32_t currd);
+    uint32_t partition(uint32_t startIndex, uint32_t endIndex, uint32_t currd, float *pivotVal);
+    void buildTree(KDNode **node, uint32_t startIndex, uint32_t endIndex, uint32_t currd, int nCores);
     void destroyNode(KDNode* node);
 };
